many updates

+++ CorneliusCup.Golf.API/Entities/Tee.cs
@@ -14,7 +14,7 @@ namespace CorneliusCup.Golf.API.Entities
 
         public int SSS { get; set; }
 
-        public int CourseRating { get; set; }
+        public decimal CourseRating { get; set; }
 
         public int SlopeRating { get; set; }
