+++ CorneliusCup.Golf.API/Responses/TeeResponse.cs
@@ -12,7 +12,7 @@ namespace CorneliusCup.Golf.API.Responses
 
         public int SSS { get; set; }
 
-        public int CourseRating { get; set; }
+        public decimal CourseRating { get; set; }
 
         public int SlopeRating { get; set; }
